# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ErrorResponse": {
				"properties": {
					"error": {
						"properties": {
							"details": {},
							"message": {
								"type": "string"
							},
							"code": {
								"type": "string"
							}
						},
						"required": [
							"message",
							"code"
						],
						"type": "object"
					}
				},
				"required": [
					"error"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"CleanerBookingListItem": {
				"properties": {
					"id": {
						"type": "string"
					},
					"service_type": {
						"type": "string"
					},
					"date": {
						"type": "string"
					},
					"time": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"customer": {
						"properties": {
							"rating": {
								"type": "number",
								"format": "double",
								"nullable": true
							},
							"name": {
								"type": "string"
							}
						},
						"required": [
							"rating",
							"name"
						],
						"type": "object",
						"nullable": true
					}
				},
				"required": [
					"id",
					"service_type",
					"date",
					"time",
					"status",
					"customer"
				],
				"type": "object",
				"additionalProperties": true
			},
			"CleanerBookingDetail": {
				"properties": {
					"id": {
						"type": "string"
					},
					"service_type": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"date": {
						"type": "string"
					},
					"time": {
						"type": "string"
					},
					"additional_info": {
						"type": "string",
						"nullable": true
					},
					"customer": {
						"properties": {
							"completed_bookings": {
								"type": "number",
								"format": "double"
							},
							"rating": {
								"type": "number",
								"format": "double",
								"nullable": true
							},
							"name": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"completed_bookings",
							"rating",
							"name",
							"id"
						],
						"type": "object",
						"nullable": true
					},
					"address": {
						"properties": {
							"entrance_notes": {
								"type": "string",
								"nullable": true
							},
							"county": {
								"type": "string",
								"nullable": true
							},
							"door_number": {
								"type": "string",
								"nullable": true
							},
							"floor_number": {
								"type": "string",
								"nullable": true
							},
							"street": {
								"type": "string"
							},
							"label": {
								"type": "string",
								"nullable": true
							}
						},
						"required": [
							"entrance_notes",
							"county",
							"door_number",
							"floor_number",
							"street",
							"label"
						],
						"type": "object",
						"nullable": true
					}
				},
				"required": [
					"id",
					"service_type",
					"status",
					"date",
					"time",
					"additional_info",
					"customer",
					"address"
				],
				"type": "object",
				"additionalProperties": true
			},
			"RateBookingRequest": {
				"properties": {
					"rating": {
						"type": "number",
						"format": "double",
						"description": "1-5"
					},
					"review": {
						"type": "string",
						"description": "0-2000 chars"
					}
				},
				"required": [
					"rating"
				],
				"type": "object",
				"additionalProperties": true
			},
			"CleanerRateRequest": {
				"$ref": "#/components/schemas/RateBookingRequest"
			},
...
			"CreateBookingRequest": {
				"properties": {
					"address_id": {
						"type": "string"
					},
					"service_type": {
						"$ref": "#/components/schemas/ServiceType"
					},
					"condition": {
						"$ref": "#/components/schemas/PropertyCondition"
					},
					"property_type": {
						"$ref": "#/components/schemas/PropertyType"
					},
					"total_sq_metres": {
						"type": "number",
						"format": "double"
					},
					"rooms": {
						"type": "number",
						"format": "double",
						"description": "1-50"
					},
					"floors": {
						"type": "number",
						"format": "double",
						"description": "1-20"
					},
					"bathrooms": {
						"type": "number",
						"format": "double",
						"description": "0-20"
					},
					"add_ons": {
						"items": {
							"$ref": "#/components/schemas/BookingAddOn"
						},
						"type": "array"
					},
					"date": {
						"type": "string",
						"description": "ISO date, must be in the future"
					},
					"time": {
						"type": "string",
						"description": "HH:MM format"
					},
					"additional_info": {
						"type": "string",
						"description": "0-2000 chars"
					},
					"use_same_cleaner": {
						"type": "boolean"
					}
				},
				"required": [
					"address_id",
					"service_type",
					"condition",
					"property_type",
					"total_sq_metres",
					"rooms",
					"floors",
					"bathrooms",
					"date",
					"time"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"ListResponse_BookingListItem_": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/BookingListItem"
						},
						"type": "array"
					},
					"meta": {
						"$ref": "#/components/schemas/PaginationMeta"
					}
				},
				"required": [
					"data",
					"meta"
				],
				"type": "object",
				"additionalProperties": true
			},
			"BookingDetail": {
				"properties": {
					"id": {
						"type": "string"
					},
					"service_type": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"date": {
						"type": "string"
					},
					"time": {
						"type": "string"
					},
					"additional_info": {
						"type": "string",
						"nullable": true
					},
					"use_same_cleaner": {
						"type": "boolean"
					},
					"cleaner": {
						"properties": {
							"location": {
								"type": "string",
								"nullable": true
							},
							"completed_bookings": {
								"type": "number",
								"format": "double"
							},
							"rating": {
								"type": "number",
								"format": "double",
								"nullable": true
							},
							"name": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"location",
							"completed_bookings",
							"rating",
							"name",
							"id"
						],
						"type": "object",
						"nullable": true
					},
					"payment": {
						"properties": {
							"status": {
								"type": "string"
							},
							"method": {
								"type": "string",
								"nullable": true
							},
							"total": {
								"type": "string",
								"nullable": true
							},
							"discount": {
								"type": "string",
								"nullable": true
							},
							"charges": {
								"type": "string",
								"nullable": true
							},
							"booking_fee": {
								"type": "string",
								"nullable": true
							}
						},
						"required": [
							"status",
							"method",
							"total",
							"discount",
							"charges",
							"booking_fee"
						],
						"type": "object"
					}
				},
				"required": [
					"id",
					"service_type",
					"status",
					"date",
					"time",
					"additional_info",
					"use_same_cleaner",
					"cleaner",
					"payment"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"AdminListResponse_Record_string.unknown__": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/Record_string.unknown_"
						},
						"type": "array"
					},
					"stats": {
						"$ref": "#/components/schemas/Record_string.unknown_"
					},
					"meta": {
						"$ref": "#/components/schemas/PaginationMeta"
					}
				},
				"required": [
					"data",
					"stats",
					"meta"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"AdminCleanerDetail": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string",
						"nullable": true
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"joined": {
						"type": "string"
					},
					"reviews": {
						"items": {
							"$ref": "#/components/schemas/ReviewSnippet"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"phone",
					"rating",
					"joined",
					"reviews"
				],
				"type": "object",
				"additionalProperties": true
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "fada-lustre-api",
		"version": "1.0.0",
		"description": "Backend API for Fada Lustre — a professional cleaning marketplace based in Derby, UK.",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Fada Lustre",
			"email": "marketing@fadalustre.com"
		}
	},
	"paths": {
...
		"/v1/cleaner/bookings": {
			"get": {
				"operationId": "ListCleanerBookings",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"total": {
											"type": "number",
											"format": "double"
										},
										"data": {
											"items": {
												"$ref": "#/components/schemas/CleanerBookingListItem"
											},
											"type": "array"
										}
									},
									"required": [
										"total",
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "List bookings assigned to the authenticated cleaner.\nSupports filtering by status (upcoming, ongoing, completed, all).",
				"summary": "List my assigned bookings",
				"tags": [
					"Cleaner Bookings"
				],
				"security": [
					{
						"jwt": [
							"cleaner"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "filter",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/cleaner/bookings/{id}": {
			"get": {
				"operationId": "GetDetail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CleanerBookingDetail"
								}
							}
						}
					},
					"403": {
						"description": "Not assigned to this booking",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Booking not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Retrieve detailed booking information including customer details,\naddress with entrance notes, and service requirements.",
				"summary": "Get booking details",
				"tags": [
					"Cleaner Bookings"
				],
				"security": [
					{
						"jwt": [
							"cleaner"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/cleaner/bookings/{id}/start": {
			"post": {
				"operationId": "Start",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"status": {
											"type": "string"
										},
										"id": {
											"type": "string"
										}
									},
									"required": [
										"status",
										"id"
									],
									"type": "object"
								}
							}
						}
					},
					"403": {
						"description": "Not assigned",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"409": {
						"description": "Invalid state transition",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Mark a scheduled booking as started. Transitions status to 'ongoing'.\nOnly the assigned cleaner can start their bookings.",
				"summary": "Start booking",
				"tags": [
					"Cleaner Bookings"
				],
				"security": [
					{
						"jwt": [
							"cleaner"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/cleaner/bookings/{id}/finish": {
			"post": {
				"operationId": "Finish",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"status": {
											"type": "string"
										},
										"id": {
											"type": "string"
										}
									},
									"required": [
										"status",
										"id"
									],
									"type": "object"
								}
							}
						}
					},
					"403": {
						"description": "Not assigned",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"409": {
						"description": "Invalid state transition",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Mark an ongoing booking as finished. Transitions status to 'done'.\nTriggers earnings calculation for the cleaner.",
				"summary": "Finish booking",
				"tags": [
					"Cleaner Bookings"
				],
				"security": [
					{
						"jwt": [
							"cleaner"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/cleaner/bookings/{id}/cancel": {
			"post": {
				"operationId": "CancelCleanerBooking",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"status": {
											"type": "string"
										},
										"id": {
											"type": "string"
										}
									},
									"required": [
										"status",
										"id"
									],
									"type": "object"
								}
							}
						}
					},
					"403": {
						"description": "Not assigned",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"409": {
						"description": "Invalid state transition",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Cancel an assigned booking from the cleaner side.\nThe booking returns to 'unassigned' for reassignment.",
				"summary": "Cancel booking (cleaner)",
				"tags": [
					"Cleaner Bookings"
				],
				"security": [
					{
						"jwt": [
							"cleaner"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/cleaner/bookings/{id}/rate": {
			"post": {
				"operationId": "RateCleanerBooking",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"rating": {
											"type": "number",
											"format": "double"
										},
										"id": {
											"type": "string"
										}
									},
									"required": [
										"rating",
										"id"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Invalid rating",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"409": {
						"description": "Already rated or not done",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Rate the customer after a completed booking.\nRating must be 1-5. Only allowed for bookings with status 'done'.",
				"summary": "Rate customer",
				"tags": [
					"Cleaner Bookings"
				],
				"security": [
					{
						"jwt": [
							"cleaner"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CleanerRateRequest"
							}
						}
					}
				}
			}
		},
...
		"/v1/bookings": {
			"post": {
				"operationId": "CreateBooking",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"price": {
											"type": "string",
											"nullable": true
										},
										"status": {
											"type": "string"
										},
										"id": {
											"type": "string"
										}
									},
									"required": [
										"price",
										"status",
										"id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Create a new cleaning booking. Requires an address, service type, property details,\nand preferred date/time. The booking starts in 'unassigned' status.",
				"summary": "Create booking",
				"tags": [
					"Bookings"
				],
				"security": [
					{
						"jwt": [
							"customer"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateBookingRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "ListBookings",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ListResponse_BookingListItem_"
								}
							}
						}
					}
				},
				"description": "List the authenticated customer's bookings with optional status filter\nand pagination. Returns bookings sorted by scheduled date.",
				"summary": "List my bookings",
				"tags": [
					"Bookings"
				],
				"security": [
					{
						"jwt": [
							"customer"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/bookings/{id}": {
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BookingDetail"
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Retrieve detailed information about a specific booking including\ncleaner info, payment status, and service details.",
				"summary": "Get booking details",
				"tags": [
					"Bookings"
				],
				"security": [
					{
						"jwt": [
							"customer"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/v1/bookings/{id}/cancel": {
			"post": {
				"operationId": "CancelBooking",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"refund_status": {
											"type": "string"
										},
										"status": {
											"type": "string"
										},
										"id": {
											"type": "string"
										}
									},
									"required": [
										"refund_status",
										"status",
										"id"
									],
									"type": "object"
								}
							}
						}
					},
					"409": {
						"description": "State conflict",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Cancel a booking. Refund eligibility depends on the cancellation timing\nrelative to the scheduled date.",
				"summary": "Cancel booking",
				"tags": [
					"Bookings"
				],
				"security": [
					{
						"jwt": [
							"customer"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/v1/bookings/{id}/rate": {
			"post": {
				"operationId": "RateBooking",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"rating": {
											"type": "number",
											"format": "double"
										},
										"id": {
											"type": "string"
										}
									},
									"required": [
										"rating",
										"id"
									],
									"type": "object"
								}
							}
						}
					},
					"409": {
						"description": "Already rated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Rate and review the cleaner after a completed booking.\nRating must be 1-5. Only allowed for bookings with status 'done'.",
				"summary": "Rate booking",
				"tags": [
					"Bookings"
				],
				"security": [
					{
						"jwt": [
							"customer"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RateBookingRequest"
							}
						}
					}
				}
			}
		},
...
		"/v1/admin/customers/{id}/book": {
			"post": {
				"operationId": "BookForCustomer",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"status": {
											"type": "string"
										},
										"price": {
											"type": "number",
											"format": "double"
										},
										"id": {
											"type": "string"
										}
									},
									"required": [
										"status",
										"price",
										"id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Create a booking on behalf of a specific customer.",
				"summary": "Book for customer",
				"tags": [
					"Admin Customers"
				],
				"security": [
					{
						"jwt": [
							"admin:customers"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AdminCreateBookingRequest"
							}
						}
					}
				}
			}
		},
...
		"/v1/admin/cleaners": {
			"get": {
				"operationId": "ListCleaners",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdminListResponse_Record_string.unknown__"
								}
							}
						}
					}
				},
				"description": "List all cleaners with filters for status, location, service type,\nand text search. Includes aggregate stats.",
				"summary": "List cleaners",
				"tags": [
					"Admin Cleaners"
				],
				"security": [
					{
						"jwt": [
							"admin:cleaners"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "location",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "service",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/admin/cleaners/{id}": {
			"get": {
				"operationId": "GetCleaner",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdminCleanerDetail"
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Retrieve detailed cleaner information including reviews, availability,\npayment history, and performance metrics.",
				"summary": "Get cleaner details",
				"tags": [
					"Admin Cleaners"
				],
				"security": [
					{
						"jwt": [
							"admin:cleaners"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/v1/admin/bookings": {
			"get": {
				"operationId": "AdminListBookings",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdminListResponse_AdminBookingListItem_"
								}
							}
						}
					}
				},
				"description": "List all bookings with filters for date, location, service type, and status.\nIncludes aggregate stats (total, scheduled, completed, cancelled).",
				"summary": "List all bookings",
				"tags": [
					"Admin Bookings"
				],
				"security": [
					{
						"jwt": [
							"admin:bookings"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "date",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "location",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "service",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "AdminCreateBooking",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"status": {
											"type": "string"
										},
										"price": {
											"type": "number",
											"format": "double"
										},
										"id": {
											"type": "string"
										}
									},
									"required": [
										"status",
										"price",
										"id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Create a booking on behalf of a customer. Used when admins receive\nphone or in-person booking requests.",
				"summary": "Create booking (admin)",
				"tags": [
					"Admin Bookings"
				],
				"security": [
					{
						"jwt": [
							"admin:bookings"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AdminCreateBookingRequest"
							}
						}
					}
				}
			}
		},
		"/v1/admin/bookings/{id}": {
			"get": {
				"operationId": "GetBooking",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdminBookingDetail"
								}
							}
						}
					},
					"404": {
						"description": "Not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Retrieve full booking details including customer/cleaner info,\nproperty details, payment status, and uploaded images.",
				"summary": "Get booking details",
				"tags": [
					"Admin Bookings"
				],
				"security": [
					{
						"jwt": [
							"admin:bookings"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/v1/admin/bookings/{id}/cancel": {
			"post": {
				"operationId": "AdminCancelBooking",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"notifications_sent": {
											"items": {
												"type": "string"
											},
											"type": "array"
										},
										"status": {
											"type": "string"
										},
										"id": {
											"type": "string"
										}
									},
									"required": [
										"notifications_sent",
										"status",
										"id"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Invalid state",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Cancel a booking and notify the customer and assigned cleaner.",
				"summary": "Cancel booking (admin)",
				"tags": [
					"Admin Bookings"
				],
				"security": [
					{
						"jwt": [
							"admin:bookings"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}